调整触及攻击的斜向距离判定

+++ src/npc_attack.cpp
@@ -223,7 +223,8 @@ void npc_attack_melee::use( npc &source, const tripoint &location ) const
         return;
     }
     if( !source.is_adjacent( critter, true ) ) {
-        if( rl_dist_exact( source.pos(), location ) <= weapon.reach_range( source ) ) {
+        if( static_cast<int>( rl_dist_exact( source.pos(), location ) ) <=
+            weapon.reach_range( source ) ) {
             add_msg_debug( debugmode::debug_filter::DF_NPC, "%s is attempting a reach attack",
                            source.disp_name() );
             // check for friendlies in the line of fire
